+++ src/main/java/core/net/OnlineWorker.java
@@ -1277,4 +1277,13 @@ public class OnlineWorker {
 		return XMLLeagueDetailsParser.parseLeagueDetails(leagueDetailsXml);
 	}
 
+	private static final Map<Integer, Map<String, TeamStats>> seriesDetailsCache = new HashMap<>();
+
+	public static Map<String, TeamStats> getSeriesDetails(int seriesId) {
+		if (!seriesDetailsCache.containsKey(seriesId)) {
+			var teamStatsMap = OnlineWorker.downloadLeagueDetails(seriesId);
+			seriesDetailsCache.put(seriesId, teamStatsMap);
+		}
+		return seriesDetailsCache.get(seriesId);
+	}
 }
